This is a employee attendance web page
The mobile attendance system has been built to eliminate the time and effort wasted in taking attendances in schools and colleges. It also greatly reduces the amount of paper resources needed in attendance data management.This is an android mobile app. It’s built to be used for school/college faculty so that they may take student attendance on their phones.
remaining is this web page is not convert into a mobile app